[Deserialization] Fix error when typealias required by protocol refers to type in @_implementationOnly module
In the added test case, the `typealias` refers to the `HiddenStruct`
type in the private module, which is imported as `@_implementationOnly`.
Because the import is `@_implementationOnly`, during deserialization, we
don’t import the private module and hence any reference to the
`HiddenStruct` type fails. In the common deserialization code path, this
causes us to skip over the `typealias` member. However, when creating
the protocol conformance, we assume that we can resolve the type to
which the `typealias` refers and thus we are crashing.

If `LangOpts.EnableDeserializationRecovery` is set to `true`, we should
do our best to recover from such failures so this patch makes the
deserialization failure handling more graceful and resolve the
right-hand side of the `typealias` as an `ErrorType`.

Fixes rdar://72891807
rdar://78035645