[Serialization] Restrict loading swiftmodule files to the builder's SDK
Serialize the SDK canonical name used when building the swiftmodule file
and use it to ensure that the swiftmodule file is loaded only with the
same SDK.

This will report unsupported configurations like:

- Installing roots between incompatible SDKs without deleting the
swiftmodule files.
- Having multiple targets in the same project using different SDKs.
- Loading a swiftmodule created with a newer SDK (and stdlib) with an
older SDK.

All of these lead to hard to investigate deserialization failures and
this change should detect them early, before reaching a deserialization
failure.

rdar://78048939