Serialization: Stop reading the control block if the format version doesn't match
One compiler may try to read from a swiftmodule from a branching
compiler with a different control block structure. We cannot rely
on that format to remain stable in a branching context.

Let’s make the compiler quit early when the swiftmodule format
doesn’t match as we do for mismatches of compiler tags.

rdar://128551774