Serialization: Support bigger buckets in the nested types lookup table
This table is used to find nested types by name. The key and hash function
are suboptimal in practice as nested types tend to have the same name.
A high number of nested types with the same name end up in the same bucket.
This leads to large buckets than can break when reading and likely
slow performance at lookup.

For now, let's only grow the maximum size of the bucket as a very safe
way to support modules with about ~10k types and above. This increases
the size of the Stdlib swiftmodulbe by less than 1 kb.

rdar://130019293